update mot so loi ve dang nhap, cap nhap them phan xoa user ,.... cap nhap tin nhan (chua realtime)

+++ app/Http/Controllers/SpecializedController.php
@@ -15,8 +15,17 @@ class SpecializedController extends Controller
      */
     public function index()
     {
+        if (Session()->has('login') && Session()->get('login') == false) {
+        # code...
+        return redirect()->route('loginView');
+        }
+        else if (!Session()->has('login')) {
+            return redirect()->route('loginView');
+        }
+        else {
         $specializeds = Specialized::all();
         return view('SpecializedManager',compact('specializeds'));
+        }
     }
     /**
      * Show the form for creating a new resource.
@@ -25,45 +34,54 @@ class SpecializedController extends Controller
      */
     public function create(Request $request)
     {
-        $specializedName = $request->specialized_name;
-        $specializedAcronym = $request->specialized_acronym;
-        $specializedDescription = $request->specialized_description;
+        if (Session()->has('login') && Session()->get('login') == false) {
+        # code...
+        return redirect()->route('loginView');
+        }
+        else if (!Session()->has('login')) {
+            return redirect()->route('loginView');
+        }
+        else {
+            $specializedName = $request->specialized_name;
+            $specializedAcronym = $request->specialized_acronym;
+            $specializedDescription = $request->specialized_description;
 
-        $specializeds = Specialized::all();
-        $kt = 0;
+            $specializeds = Specialized::all();
+            $kt = 0;
 
-        foreach ($specializeds as $specialized) {
-            if ($specialized->specialized_name == $specializedName) {
-                # code...
-                $kt = 1;
+            foreach ($specializeds as $specialized) {
+                if ($specialized->specialized_name == $specializedName) {
+                    # code...
+                    $kt = 1;
+                }
             }
-        }
 
-        if ($specializedName == "") {
-            $error = 'Tên khoa không được bỏ trống';
-        }
-        else if ($specializedAcronym == "") {
-            $error = 'Viết tắt khoa không được bỏ trống';
-        }  
-        else if ($specializedDescription == "") {
-            $error = 'Mô tả không được bỏ trống';
-        }
-        else if ($kt == 1) {
-            $error = 'Tên khoa đã tồn tại';
-        }else{
-            $SpecializedNew = new Specialized;
-            $SpecializedNew->timestamps = false;
-            $SpecializedNew->specialized_name=$specializedName;
-            $SpecializedNew->specialized_acronym=$specializedAcronym;
-            $SpecializedNew->specialized_description=$specializedDescription;
-            $SpecializedNew->save();
+            if ($specializedName == "") {
+                $error = 'Tên khoa không được bỏ trống';
+            }
+            else if ($specializedAcronym == "") {
+                $error = 'Viết tắt khoa không được bỏ trống';
+            }  
+            else if ($specializedDescription == "") {
+                $error = 'Mô tả không được bỏ trống';
+            }
+            else if ($kt == 1) {
+                $error = 'Tên khoa đã tồn tại';
+            }else{
+                $SpecializedNew = new Specialized;
+                $SpecializedNew->timestamps = false;
+                $SpecializedNew->specialized_name=$specializedName;
+                $SpecializedNew->specialized_acronym=$specializedAcronym;
+                $SpecializedNew->specialized_description=$specializedDescription;
+                $SpecializedNew->save();
 
-            //quay tro lai giao dien voi thong bao thanh cong
-            $request->session()->put('notice', 'thêm thành công');
+                //quay tro lai giao dien voi thong bao thanh cong
+                $request->session()->put('notice', 'thêm thành công');
+                return redirect()->route('SpecializedManagementView');
+            }
+            $request->session()->put('notice',$error);
             return redirect()->route('SpecializedManagementView');
         }
-        $request->session()->put('notice',$error);
-        return redirect()->route('SpecializedManagementView');
     }
 
     /**
@@ -85,8 +103,17 @@ class SpecializedController extends Controller
      */
     public function show(Request $request)
     {
-        $specializeds = Specialized::Where('specialized_Name','like','%'.$request->name.'%')->get();
-        return view('SpecializedManager',compact('specializeds'));
+        if (Session()->has('login') && Session()->get('login') == false) {
+        # code...
+        return redirect()->route('loginView');
+        }
+        else if (!Session()->has('login')) {
+            return redirect()->route('loginView');
+        }
+        else {
+            $specializeds = Specialized::Where('specialized_Name','like','%'.$request->name.'%')->get();
+            return view('SpecializedManager',compact('specializeds'));
+        }
     }
 
     /**
@@ -97,40 +124,49 @@ class SpecializedController extends Controller
      */
     public function edit(Request $request)
     {
-        $specializedName = $request->specialized_name;
-        $specializedAcronym = $request->specialized_acronym;
-        $specializedDescription = $request->specialized_description;
+        if (Session()->has('login') && Session()->get('login') == false) {
+        # code...
+        return redirect()->route('loginView');
+        }
+        else if (!Session()->has('login')) {
+            return redirect()->route('loginView');
+        }
+        else {
+            $specializedName = $request->specialized_name;
+            $specializedAcronym = $request->specialized_acronym;
+            $specializedDescription = $request->specialized_description;
 
-        $id = Specialized::where('specialized_name','=',$specializedName)->get('specialized_id');
+            $id = Specialized::where('specialized_name','=',$specializedName)->get('specialized_id');
 
-        $specializeds = Specialized::all();
-        $kt = 0;
+            $specializeds = Specialized::all();
+            $kt = 0;
 
-        foreach ($specializeds as $specialized) {
-            if ($specialized->specialized_name == $specializedName) {
-                # code...
-                $kt = 1;
+            foreach ($specializeds as $specialized) {
+                if ($specialized->specialized_name == $specializedName) {
+                    # code...
+                    $kt = 1;
+                }
             }
-        }
 
-        if ($specializedName == "") {
-            $error = 'Tên khoa không được bỏ trống';
-        }
-        else if ($specializedAcronym == "") {
-            $error = 'Viết tắt khoa không được bỏ trống';
-        }  
-        else if ($specializedDescription == "") {
-            $error = 'Mô tả không được bỏ trống';
-        }
-        else if ($kt == 0) {
-            $error = 'Tên khoa đã tồn tại';
-        }else{
-            $specializedNew = Specialized::where('specialized_id','=',$id[0]['specialized_id'])->update(['specialized_name'=>$specializedName,'specialized_acronym'=>$specializedAcronym,'specialized_description'=>$specializedDescription,]);
-            $request->session()->put('notice', 'Sửa thành công');
+            if ($specializedName == "") {
+                $error = 'Tên khoa không được bỏ trống';
+            }
+            else if ($specializedAcronym == "") {
+                $error = 'Viết tắt khoa không được bỏ trống';
+            }  
+            else if ($specializedDescription == "") {
+                $error = 'Mô tả không được bỏ trống';
+            }
+            else if ($kt == 0) {
+                $error = 'Tên khoa đã tồn tại';
+            }else{
+                $specializedNew = Specialized::where('specialized_id','=',$id[0]['specialized_id'])->update(['specialized_name'=>$specializedName,'specialized_acronym'=>$specializedAcronym,'specialized_description'=>$specializedDescription,]);
+                $request->session()->put('notice', 'Sửa thành công');
+                return redirect()->route('SpecializedManagementView');
+            }
+            $request->session()->put('notice',$error);
             return redirect()->route('SpecializedManagementView');
         }
-        $request->session()->put('notice',$error);
-        return redirect()->route('SpecializedManagementView');
     }
 
     /**
@@ -153,8 +189,17 @@ class SpecializedController extends Controller
      */
     public function destroy(Request $request, $id)
     {
-        $specializeds = Specialized::where('specialized_id','=',$id)->delete();
-        $request->session()->put('notice', 'Xóa thành công');
-        return redirect()->route('SpecializedManagementView');
+        if (Session()->has('login') && Session()->get('login') == false) {
+        # code...
+        return redirect()->route('loginView');
+        }
+        else if (!Session()->has('login')) {
+            return redirect()->route('loginView');
+        }
+        else {
+            $specializeds = Specialized::where('specialized_id','=',$id)->delete();
+            $request->session()->put('notice', 'Xóa thành công');
+            return redirect()->route('SpecializedManagementView');
+        }
     }
 }
