+++ app/Http/Controllers/SubjectController.php
@@ -15,9 +15,17 @@ class SubjectController extends Controller
     public function index()
     {
         //
-        $subjects = Subject::all();
-        return view('subjectManagement',compact('subjects'));
-
+        if (Session()->has('login') && Session()->get('login') == false) {
+        # code...
+        return redirect()->route('loginView');
+        }
+        else if (!Session()->has('login')) {
+            return redirect()->route('loginView');
+        }
+        else {
+            $subjects = Subject::all();
+            return view('subjectManagement',compact('subjects'));
+        }
     }
 
     /**
@@ -28,40 +36,49 @@ class SubjectController extends Controller
     public function create(Request $request)
     {
         //
-        $subjectName = $request->subject_name;
-        $numberCredit = $request->number_credit;
+        if (Session()->has('login') && Session()->get('login') == false) {
+        # code...
+        return redirect()->route('loginView');
+        }
+        else if (!Session()->has('login')) {
+            return redirect()->route('loginView');
+        }
+        else {
+            $subjectName = $request->subject_name;
+            $numberCredit = $request->number_credit;
 
-        $subjects = Subject::all();
-        $kt = 0;
+            $subjects = Subject::all();
+            $kt = 0;
 
-        foreach ($subjects as $subject) {
-            if ($subject->subject_name == $subjectName) {
-                # code...
-                $kt = 1;
+            foreach ($subjects as $subject) {
+                if ($subject->subject_name == $subjectName) {
+                    # code...
+                    $kt = 1;
+                }
             }
-        }
 
-        if ($subjectName == "") {
-            $error = 'Tên môn không được bỏ trống';
-        }
-        else if ($numberCredit == "") {
-            $error = 'số tín chỉ không được bỏ trống';
-        }  
-        else if ($kt !=0) {
-            $error = 'Tên môn đã tồn tại';
-        }else{
-            $subjectNew = new Subject;
-            $subjectNew->timestamps = false;
-            $subjectNew->subject_name=$subjectName;
-            $subjectNew->subject_numberCredit=$numberCredit;
-            $subjectNew->save();
+            if ($subjectName == "") {
+                $error = 'Tên môn không được bỏ trống';
+            }
+            else if ($numberCredit == "") {
+                $error = 'số tín chỉ không được bỏ trống';
+            }  
+            else if ($kt !=0) {
+                $error = 'Tên môn đã tồn tại';
+            }else{
+                $subjectNew = new Subject;
+                $subjectNew->timestamps = false;
+                $subjectNew->subject_name=$subjectName;
+                $subjectNew->subject_numberCredit=$numberCredit;
+                $subjectNew->save();
 
-            //quay tro lai giao dien voi thong bao thanh cong
-            $request->session()->put('notice', 'thêm thành công');
+                //quay tro lai giao dien voi thong bao thanh cong
+                $request->session()->put('notice', 'thêm thành công');
+                return redirect()->route('SubjectMangementView');
+            }
+            $request->session()->put('notice',$error);
             return redirect()->route('SubjectMangementView');
         }
-        $request->session()->put('notice',$error);
-        return redirect()->route('SubjectMangementView');
     }
 
     /**
@@ -84,8 +101,17 @@ class SubjectController extends Controller
     public function show(Request $request)
     {
         //
-        $subjects = Subject::Where('subject_name','like','%'.$request->name.'%')->get();
-        return view('subjectManagement',compact('subjects'));
+        if (Session()->has('login') && Session()->get('login') == false) {
+        # code...
+        return redirect()->route('loginView');
+        }
+        else if (!Session()->has('login')) {
+            return redirect()->route('loginView');
+        }
+        else {
+            $subjects = Subject::Where('subject_name','like','%'.$request->name.'%')->get();
+            return view('subjectManagement',compact('subjects'));
+        }
     }
 
     /**
@@ -97,36 +123,45 @@ class SubjectController extends Controller
     public function edit(Request $request, $id)
     {
         //
-        $subjectName = $request->subject_name;
-        $numberCredit = $request->number_credit;
+        if (Session()->has('login') && Session()->get('login') == false) {
+        # code...
+        return redirect()->route('loginView');
+        }
+        else if (!Session()->has('login')) {
+            return redirect()->route('loginView');
+        }
+        else {
+            $subjectName = $request->subject_name;
+            $numberCredit = $request->number_credit;
 
-        //$id = Subject::where('subject_name','=',$subjectName)->get('subject_id');
+            //$id = Subject::where('subject_name','=',$subjectName)->get('subject_id');
 
-        $subjects = Subject::all();
-        $kt = 0;
+            $subjects = Subject::all();
+            $kt = 0;
 
-        foreach ($subjects as $subject) {
-            if ($subject->subject_name == $subjectName) {
-                # code...
-                $kt = 1;
+            foreach ($subjects as $subject) {
+                if ($subject->subject_name == $subjectName) {
+                    # code...
+                    $kt = 1;
+                }
             }
-        }
 
-        if ($subjectName == "") {
-            $error = 'Tên môn không được bỏ trống';
-        }
-        else if ($numberCredit == "") {
-            $error = 'Số tín chỉ không được bỏ trống';
-        }  
-        else if ($kt != 0) {
-            $error = 'Tên môn đã tồn tại';
-        }else{
-            $subjectNew = Subject::where('subject_id','=',$ids)->update(['subject_name'=>$subjectName,'subject_numberCredit'=>$numberCredit,]);
-            $request->session()->put('notice', 'Sửa thành công');
+            if ($subjectName == "") {
+                $error = 'Tên môn không được bỏ trống';
+            }
+            else if ($numberCredit == "") {
+                $error = 'Số tín chỉ không được bỏ trống';
+            }  
+            else if ($kt != 0) {
+                $error = 'Tên môn đã tồn tại';
+            }else{
+                $subjectNew = Subject::where('subject_id','=',$ids)->update(['subject_name'=>$subjectName,'subject_numberCredit'=>$numberCredit,]);
+                $request->session()->put('notice', 'Sửa thành công');
+                return redirect()->route('SubjectMangementView');
+            }
+            $request->session()->put('notice',$error);
             return redirect()->route('SubjectMangementView');
         }
-        $request->session()->put('notice',$error);
-        return redirect()->route('SubjectMangementView');
     }
 
     /**
@@ -150,8 +185,17 @@ class SubjectController extends Controller
     public function destroy(Request $request, $id)
     {
         //
-        $subjects = Subject::where('subject_id','=',$id)->delete();
-        $request->session()->put('notice', 'Xóa thành công');
-        return redirect()->route('SubjectMangementView');
+        if (Session()->has('login') && Session()->get('login') == false) {
+        # code...
+        return redirect()->route('loginView');
+        }
+        else if (!Session()->has('login')) {
+            return redirect()->route('loginView');
+        }
+        else {
+            $subjects = Subject::where('subject_id','=',$id)->delete();
+            $request->session()->put('notice', 'Xóa thành công');
+            return redirect()->route('SubjectMangementView');
+        }
     }
 }
